feat: use beamer-like itemize item and subitem templates
This brings the itemize templates more in line with beamer. The changes
are mostly that the symbols are slightly raised. Alignment with te
subsubitem symbol is not great, but I will defer it to another update.
This also closes #49, since we now rely on `$...$` instead of `\(...\)`.

fix: avoid coloring bullets based on math colors