fix: enable background macro for alternative color themes
Add a macro to override the colors for dark and light versions of the
theme, which can use arbitrary `\setbeamercolor` calls to setup the dark
and light versions however one may choose. Thanks to @ohnemax for
raising the issue and suggesting a fix in the first place.

feat: add hook system for managing dark and light themes